tune AES in Android, both armv8, armv7 and armeabi

+++ Core/MMKVMetaInfo.hpp
@@ -46,11 +46,11 @@ struct MMKVMetaInfo {
     uint32_t m_version = MMKVVersionSequence;
     uint32_t m_sequence = 0; // full write-back count
     unsigned char m_vector[AES_KEY_LEN] = {};
-    size_t m_actualSize = 0;
+    uint32_t m_actualSize = 0;
 
     // confirmed info: it's been synced to file
     struct {
-        size_t lastActualSize = 0;
+        uint32_t lastActualSize = 0;
         uint32_t lastCRCDigest = 0;
         uint32_t __reserved__[16] = {};
     } m_lastConfirmedMetaInfo;
